Simplifying test bench... weird issue all mems empty though..
